fork: store decoded samples as i16 instead of f32
xsynth's original design holds every decoded sample as f32 stereo at the
output rate. For WörliTzer's 452-sample library that's ~950 MB resident;
two heavy SFZ tracks (Wurlitzer + Fender Bass = ~1.5 GB) exceed Move's
1.85 GB RAM and Rust's alloc handler aborts the shim.

Switching internal storage to i16 halves the footprint while keeping
16-bit dynamic range (most input samples are 16-bit WAVs anyway). Per
sample read we now convert i16→f32 with one multiply by 1/32768 — a
cost so small it disappears next to interpolation + envelope + filter.

Touchpoints:
- ProcessedSample + spawner_params_list + voice_spawners type changes
  from Arc<[Arc<[f32]>]> to Arc<[Arc<[i16]>]>.
- resample_vecs / resample_vec converts the f32 resampler output to i16
  before returning. Identity-rate samples skip the resampler entirely.
- F32BufferSampler renamed I16BufferSampler; get() reads i16 and scales
  to f32. BufferSamplers enum variant updated.
- SF2 sample path: pass-through case converts f32 decoder output to i16
  with the same clamp-and-cast as the resample path.

Net effect on Move: WörliTzer (~475 MB) + Fender Bass (~300 MB) load
together within budget.

+++ core/src/soundfont/audio.rs
@@ -26,7 +26,7 @@ pub enum AudioLoadError {
     NoTracks(PathBuf),
 }
 
-type ProcessedSample = (Arc<[Arc<[f32]>]>, u32);
+type ProcessedSample = (Arc<[Arc<[i16]>]>, u32);
 
 pub(super) fn load_audio_file(
     path: &PathBuf,
@@ -159,7 +159,7 @@ impl BuilderVecs {
         sample_rate: f32,
         new_sample_rate: f32,
         channels: ChannelCount,
-    ) -> Arc<[Arc<[f32]>]> {
+    ) -> Arc<[Arc<[i16]>]> {
         let mut vecs = self.vecs;
 
         if channels == ChannelCount::Mono && vecs.len() >= 2 {
